server/requirements.txt
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "3.11"
anyio==4.3.0 ; python_version >= "3.10" and python_version < "3.11"
cachetools==5.3.3 ; python_version >= "3.10" and python_version < "3.11"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "3.11"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and (sys_platform == "win32" or platform_system == "Windows")
db-dtypes==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.110.0 ; python_version >= "3.10" and python_version < "3.11"
google-api-core==2.17.1 ; python_version >= "3.10" and python_version < "3.11"
google-api-core[grpc]==2.17.1 ; python_version >= "3.10" and python_version < "3.11"
google-auth-oauthlib==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
google-auth==2.28.2 ; python_version >= "3.10" and python_version < "3.11"
google-cloud-bigquery==3.19.0 ; python_version >= "3.10" and python_version < "3.11"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "3.11"
google-cloud-secret-manager==2.18.3 ; python_version >= "3.10" and python_version < "3.11"
google-cloud-storage==2.15.0 ; python_version >= "3.10" and python_version < "3.11"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "3.11"
google-resumable-media==2.7.0 ; python_version >= "3.10" and python_version < "3.11"
googleapis-common-protos==1.63.0 ; python_version >= "3.10" and python_version < "3.11"
googleapis-common-protos[grpc]==1.63.0 ; python_version >= "3.10" and python_version < "3.11"
grpc-google-iam-v1==0.13.0 ; python_version >= "3.10" and python_version < "3.11"
grpcio-status==1.60.1 ; python_version >= "3.10" and python_version < "3.11"
grpcio==1.60.1 ; python_version >= "3.10" and python_version < "3.11"
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.11"
httptools==0.6.1 ; python_version >= "3.10" and python_version < "3.11"
idna==3.6 ; python_version >= "3.10" and python_version < "3.11"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "3.11"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "3.11"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "3.11"
packaging==24.0 ; python_version >= "3.10" and python_version < "3.11"
pandas-gbq==0.22.0 ; python_version >= "3.10" and python_version < "3.11"
pandas==2.2.1 ; python_version >= "3.10" and python_version < "3.11"
proto-plus==1.23.0 ; python_version >= "3.10" and python_version < "3.11"
protobuf==4.25.3 ; python_version >= "3.10" and python_version < "3.11"
pyarrow==15.0.1 ; python_version >= "3.10" and python_version < "3.11"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "3.11"
pyasn1==0.5.1 ; python_version >= "3.10" and python_version < "3.11"
pydantic-core==2.16.3 ; python_version >= "3.10" and python_version < "3.11"
pydantic==2.6.4 ; python_version >= "3.10" and python_version < "3.11"
pydata-google-auth==1.8.2 ; python_version >= "3.10" and python_version < "3.11"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "3.11"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "3.11"
pytz==2024.1 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.11"
requests-oauthlib==1.4.0 ; python_version >= "3.10" and python_version < "3.11"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.11"
rsa==4.9 ; python_version >= "3.10" and python_version < "3.11"
scikit-learn==1.4.1.post1 ; python_version >= "3.10" and python_version < "3.11"
scipy==1.12.0 ; python_version >= "3.10" and python_version < "3.11"
setuptools==69.1.1 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "3.11"
starlette==0.36.3 ; python_version >= "3.10" and python_version < "3.11"
threadpoolctl==3.3.0 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.10.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "3.11"
urllib3==1.26.18 ; python_version >= "3.10" and python_version < "3.11"
uvicorn[standard]==0.28.0 ; python_version >= "3.10" and python_version < "3.11"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "3.11"
watchfiles==0.21.0 ; python_version >= "3.10" and python_version < "3.11"
websockets==12.0 ; python_version >= "3.10" and python_version < "3.11"
